Initial support both ASCII and UNICODE builds on Windows.
This requires using a different character type for Windows (TCHAR), as well as
a string literal wrapper (TEXT) and alternative string function names (_tcs*).

To not break POSIX and to avoid spoiling the global namespace, I've made
simple wrappers prefixed with _tinydir_ for all of these things.

This is rather ugly, admittedly.